+++ docs/pages/installation/api/constructor-options.js
@@ -51,11 +51,25 @@ export default [
                 default: '<code>3500</code>'
             },
             {
+                name: '<code>defaultSnackbarPosition</code>',
+                description: `Default snackbar <code>position</code> attribute.`,
+                type: 'Number',
+                values: '—',
+                default: '<code></code>'
+            },
+            {
                 name: '<code>defaultToastDuration</code>',
                 description: `Default toast <code>duration</code> attribute.`,
                 type: 'Number',
                 values: '—',
                 default: '<code>2000</code>'
+            },
+            {
+                name: '<code>defaultToastPosition</code>',
+                description: `Default toast <code>position</code> attribute.`,
+                type: 'Number',
+                values: '—',
+                default: '<code></code>'
             },
             {
                 name: '<code>defaultTooltipType</code>',